49395 & 49397

See merge request stfalcon/fwdays!1

+++ src/Stfalcon/Bundle/EventBundle/Command/StfalconMailerCommand.php
@@ -2,44 +2,43 @@
 
 namespace Stfalcon\Bundle\EventBundle\Command;
 
+use Application\Bundle\DefaultBundle\Service\MyMailer;
 use Symfony\Bundle\FrameworkBundle\Command\ContainerAwareCommand;
-
-use Symfony\Component\Console\Input\InputArgument,
-    Symfony\Component\Console\Input\InputInterface,
-    Symfony\Component\Console\Input\InputOption,
-    Symfony\Component\Console\Output\OutputInterface;
-
+use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Input\InputOption;
+use Symfony\Component\Console\Output\OutputInterface;
 use Stfalcon\Bundle\EventBundle\Entity\Mail;
 
 /**
- * Class StfalconMailerCommand
+ * Class StfalconMailerCommand.
  */
 class StfalconMailerCommand extends ContainerAwareCommand
 {
     /**
-     * Set options
+     * Set options.
      */
     protected function configure()
     {
         $this
             ->setName('stfalcon:mailer')
             ->setDescription('Send message from queue')
-            ->addOption('amount', null, InputOption::VALUE_OPTIONAL, 'Amount of mails which will send per operation. Default 10.');
+            ->addOption('amount', null, InputOption::VALUE_OPTIONAL, 'Amount of mails which will send per operation. Default 10.')
+            ->addOption('host', null, InputOption::VALUE_OPTIONAL, 'Site host. Default fwdays.com.');
     }
 
     /**
-     * Execute command
-     *
-     * @param InputInterface  $input  Input
-     * @param OutputInterface $output Output
+     * @param InputInterface  $input
+     * @param OutputInterface $output
      *
      * @return int|null|void
+     *
+     * @throws \Doctrine\ORM\OptimisticLockException
      */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
         /** @var \Symfony\Component\Routing\RequestContext $context */
         $context = $this->getContainer()->get('router')->getContext();
-        $context->setHost('frameworksdays.com');
+        $context->setHost('fwdays.com');
         $context->setScheme('http');
 
         $limit = 10;
@@ -48,23 +47,35 @@ class StfalconMailerCommand extends ContainerAwareCommand
             $limit = (int) $input->getOption('amount');
         }
 
+        if ($input->getOption('host')) {
+            $context->setHost($input->getOption('host'));
+        }
+
         /** @var $em \Doctrine\ORM\EntityManager */
         $em = $this->getContainer()->get('doctrine.orm.entity_manager');
-        /** @var $mailer \Swift_Mailer */
-        $mailer = $this->getContainer()->get('mailer');
+        /** @var $mailer MyMailer */
+        $mailer = $this->getContainer()->get('app.my_mailer.service');
         /** @var $mailerHelper \Stfalcon\Bundle\EventBundle\Helper\StfalconMailerHelper */
         $mailerHelper = $this->getContainer()->get('stfalcon_event.mailer_helper');
         /** @var $queueRepository \Stfalcon\Bundle\EventBundle\Repository\MailQueueRepository */
         $queueRepository = $em->getRepository('StfalconEventBundle:MailQueue');
 
         $mailsQueue = $queueRepository->getMessages($limit);
-
-        /** @var $mail Mail */
+        $logger = $this->getContainer()->get('logger');
+        /* @var $mail Mail */
         foreach ($mailsQueue as $item) {
             $user = $item->getUser();
             $mail = $item->getMail();
 
-            if (!($user && $mail) || !$user->isSubscribe()) {
+            if (!(
+                $user &&
+                $mail &&
+                $user->isEnabled() &&
+                $user->isSubscribe() &&
+                $user->isEmailExists() &&
+                filter_var($user->getEmail(), FILTER_VALIDATE_EMAIL)
+            )) {
+                $mail->decTotalMessages();
                 $em->remove($item);
                 $em->flush();
                 continue;
@@ -73,42 +84,45 @@ class StfalconMailerCommand extends ContainerAwareCommand
             try {
                 $message = $mailerHelper->formatMessage($user, $mail);
             } catch (\Exception $e) {
-                $this->getContainer()->get('logger')->addError($e->getMessage(), array('email' => $user->getEmail()));
+                $logger->addError('Mailer:'.$e->getMessage(), ['email' => $user->getEmail()]);
 
-                $mail->setTotalMessages($mail->getTotalMessages() - 1);
-                if ($mail->getSentMessages() == $mail->getTotalMessages()) {
-                    $mail->setStart(false);
-                }
-
-                $em->persist($mail);
+                $mail->decTotalMessages();
                 $em->remove($item);
                 $em->flush();
                 continue;
             }
 
-
-            //add header tag for unsubscribe
             $headers = $message->getHeaders();
-            $http = $this->getContainer()->get('router')->generate('unsubscribe',
+            $http = $this->getContainer()->get('router')->generate(
+                'unsubscribe',
                 [
                     'hash' => $user->getSalt(),
-                    'userId' => $user->getId()
-                ], true);
+                    'userId' => $user->getId(),
+                    'mailId' => $mail->getId(),
+                ],
+                true
+            );
 
-            $headers->addTextHeader('List-Unsubscribe', '<' . $http . '>');
-
-            if ($mailer->send($message)) {
-                $mail->setSentMessages($mail->getSentMessages() + 1);
+            $headers->removeAll('List-Unsubscribe');
+            $headers->addTextHeader('List-Unsubscribe', '<'.$http.'>');
+            $failed = [];
+            if ($mailer->send($message, $failed)) {
+                $mail->incSentMessage();
                 $item->setIsSent(true);
-
-                if ($mail->getSentMessages() == $mail->getTotalMessages()) {
-                    $mail->setStart(false);
-                }
-
-                $em->persist($mail);
-                $em->persist($item);
                 $em->flush();
+            } else {
+                $logger->addError('Mailer send exception', [
+                    'mail_id' => $mail->getId(),
+                    'user_id' => $user->getId(),
+                    'error_swift_message' => $failed['error_swift_message'],
+                    'error_swift_code' => $failed['error_swift_code'],
+                    'error_swift_trace' => $failed['error_swift_trace'],
+                    'error_exception_message' => $failed['error_exception_message'],
+                    'error_exception_code' => $failed['error_exception_code'],
+                    'error_exception_trace' => $failed['error_exception_trace'],
+                ]);
             }
         }
+        $em->flush();
     }
 }
